Rename to CoordinatesExt
